requirements.txt
streamlit
reportlab
qrcode[pil]
Pillow
streamlit
reportlab
qrcode[pil]
Pillow
flask
requests
streamlit
pybluez